Non-functional cleanup issues collected from recent work;
1) Recent compiler warnings in ha_innodb.cc and i_s.cc that occur
   on VS2010-64.
2) Converted some fprintf(stderr,...) mesages to ib_logf() in
   fil0fil.cc
3) Fixed innodb_bug16066351 so that it will only run on
   lower_case_table_names=0 machines instead of trying to skip
   windows and certain versions of MacOS.  It did not list MacOS 10.8
   and was failing there.
4) A message in srv/srv0space.cc Tablespace::check_file_spec() was
   not displaying the correct info.  Fixed in the code and
   in innodb-log-file.test.
5) A couple other minor cleanups. 

Approved in rb:3616 by Marko

+++ storage/innobase/handler/ha_innodb.cc
@@ -12397,7 +12397,7 @@ innodb_show_status(
 
 	ret_val= stat_print(thd, innobase_hton_name,
 				(uint) strlen(innobase_hton_name),
-				STRING_WITH_LEN(""), str, flen);
+				STRING_WITH_LEN(""), str, uint(flen));
 
 	my_free(str);
